@@ -1,30 +1,39 @@
 [package]
 name = "sps_plot"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Alex Conley <imaconley1@gmail.com>"]
 edition = "2021"
-rust-version = "1.72"
+include = ["LICENSE-APACHE", "LICENSE-MIT", "**/*.rs", "Cargo.toml"]
+rust-version = "1.88"
+
+[package.metadata.docs.rs]
+all-features = true
+targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
 
 [dependencies]
-egui = "0.27.0"
-eframe = { version = "0.27.0", default-features = false, features = [
+egui = "0.33"
+eframe = { version = "0.33", default-features = false, features = [
     "default_fonts", # Embed the default egui fonts.
     "glow",          # Use the glow rendering backend. Alternative: "wgpu".
     "persistence",   # Enable restoring app state when restarting the app.
+    "wayland",       # To support Linux (and CI)
+    "x11",           # To support older Linux distributions (restores one of the default features)
 ] }
-log = "0.4"
+log = "0.4.28"
 serde = { version = "1", features = ["derive"] }
 
-egui_plot = { version = "0.27.2", features = ["serde"] }
-egui_extras = "0.27.2"
+egui_plot = "0.34.0"
+
+egui_extras = "0.33.0"
 
 # Conditional dependencies for native and wasm targets
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-env_logger = "0.10"
-rfd = "0.13"
+env_logger = "0.11.8"
+rfd = "0.15.4"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen-futures = "0.4"
+wasm-bindgen-futures = "0.4.50"
+web-sys = "0.3.70"              # to access the DOM (to hide the loading text)
 
 [profile.release]
 opt-level = 2 # fast and small wasm
@@ -32,5 +41,3 @@ opt-level = 2 # fast and small wasm
 # Optimize all dependencies even in debug builds
 [profile.dev.package."*"]
 opt-level = 2
-
-[patch.crates-io]
